feat: rename rules.json to index.json
