+++ src/rulesets/oas/__tests__/openapi-tags.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('openapi-tags', () => {
   const s = new Spectral();
