+++ src/rulesets/oas/__tests__/operation-2xx-response.ts
@@ -1,6 +1,6 @@
 import { commonOasFunctions } from '../';
 import { Spectral } from '../../../spectral';
-import { rules } from '../rules.json';
+import { rules } from '../index.json';
 
 const { oasOp2xxResponse } = commonOasFunctions();
