+++ src/rulesets/oas/__tests__/operation-default-response.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('operation-default-response', () => {
   const s = new Spectral();
