+++ src/rulesets/oas/__tests__/path-not-include-query.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('path-not-include-query', () => {
   const s = new Spectral();
