+++ src/rulesets/oas2/__tests__/api-schemes.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('api-schemes', () => {
   const s = new Spectral();
