+++ src/rulesets/oas2/__tests__/model-description.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('model-description', () => {
   const s = new Spectral();
