+++ src/rulesets/oas3/__tests__/server-not-example.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../rules.json';
+import * as ruleset from '../index.json';
 
 describe('server-not-example.com', () => {
   const s = new Spectral();
